requirements.txt
poetry
python-decouple
